Save backup using non-destructive stash.
Update the logic to use my fork of git2-rs and libgit2, which expose
stash create and stash store operations so that using a stash to take
the backup does not change files on the file system.

This is friendlier to projects which have file watchers that react
to changes in the file system.